configure: Add --enable-migration-from-qemu-kvm
This switch will turn on all the migration compat bits needed to
perform migration from qemu-kvm to qemu. It's just a stub for now.

This compat will break incoming migration from qemu < 1.3, but for
distros where qemu-kvm was the only shipped package for years it's
not a big loss (and I don't know any way to avoid it).

Signed-off-by: Cole Robinson <crobinso@redhat.com>