acpi_piix4: Drop minimum_version_id to handle qemu-kvm migration
qemu-kvm 1.2 advertised version_id=2, but it was not the same
format as qemu.git version_id=2.

commit b0b873a07872f7ab7f66f259c73fb9dd42aa66a9 added the qemu-kvm
format to qemu.git, but was forced to call it version_id=3, and
bumped minimum_version_id to 3. This breaks incoming migration from
qemu-kvm.

If --enable-migration-from-qemu-kvm is enabled, drop minimum_version_id
to 2. Migration from qemu-kvm version_id=2 and qemu 1.3+ version_id=3
works, but migration from qemu < 1.3 is broken.

Signed-off-by: Cole Robinson <crobinso@redhat.com>